Rename .ce.vue -> .vue; no longer needed
