+++ openlibrary/components/ObservationForm/components/CardBody.vue
@@ -14,7 +14,7 @@
 </template>
 
 <script>
-import OLChip from './OLChip.ce'
+import OLChip from './OLChip'
 
 import { updateObservation } from '../ObservationService'
